Code cleanup

+++ Content/Projectiles/Weapons/Melee/Boomerangs/LeecharangProjectile.cs
@@ -1,6 +1,5 @@
 using Microsoft.Xna.Framework;
 using Terraria;
-using Terraria.ID;
 using Terraria.ModLoader;
 using TerrariaParadox.Content.Debuffs.DoT;
 using TerrariaParadox.Content.Dusts.Tiles.Blocks;
@@ -10,6 +9,8 @@ namespace TerrariaParadox.Content.Projectiles.Weapons.Melee.Boomerangs;
 
 public class LeecharangProjectile : ModdedFriendlyProjectile
 {
+    public Vector2 InitialProjVelocity;
+    public bool Returning;
     public override int Frames => 1;
     public override int AnimationDuration => 0;
     public override int Width => 24;
@@ -19,21 +20,22 @@ public class LeecharangProjectile : ModdedFriendlyProjectile
     public override bool PassThroughBlocks => false;
     public override int Pierce => 1; //so it can return after having hit an enemy
     public override float RotationHelper => 0;
-    public bool Returning = false;
-    public Vector2 InitialProjVelocity;
+
     public override void CustomAI()
     {
-        Player player = Main.player[Projectile.owner];
-        
+        var player = Main.player[Projectile.owner];
+
         AITimer1++;
 
         Projectile.rotation += AITimer1 * 0.15f;
-        
+
         if (AITimer1 % 2 == 0)
         {
-            Dust Boomer = Dust.NewDustDirect(Projectile.position, Projectile.width, Projectile.height, ModContent.DustType<AssecstoneDust>());
+            var Boomer = Dust.NewDustDirect(Projectile.position, Projectile.width, Projectile.height,
+                ModContent.DustType<AssecstoneDust>());
             Boomer.noGravity = true;
         }
+
         if (!Returning)
         {
             switch (AITimer1)
@@ -43,7 +45,7 @@ public class LeecharangProjectile : ModdedFriendlyProjectile
                     InitialProjVelocity = Projectile.velocity;
                     break;
                 }
-                case float Slowdown when (Slowdown >= 55 && Slowdown < 85): //make it slow down
+                case float Slowdown when Slowdown >= 55 && Slowdown < 85: //make it slow down
                 {
                     Projectile.velocity -= InitialProjVelocity / 30f;
                     break;
@@ -60,22 +62,23 @@ public class LeecharangProjectile : ModdedFriendlyProjectile
             //Projectile.rotation -= AITimer1 * 0.15f;
             switch (AITimer1)
             {
-                case float Speedup when (Speedup > 85 && Speedup <= 110): //make it speed up while going back towards the player
+                case float Speedup
+                    when Speedup > 85 && Speedup <= 110: //make it speed up while going back towards the player
                 {
                     AITimer2++;
-                    Projectile.velocity = ((player.Center - Projectile.Center).SafeNormalize(Vector2.Zero) * Leecharang.ShootSpeed * 1.25f) * (AITimer2 * 0.04f);
+                    Projectile.velocity = (player.Center - Projectile.Center).SafeNormalize(Vector2.Zero) *
+                                          Leecharang.ShootSpeed * 1.25f * (AITimer2 * 0.04f);
                     break;
                 }
-                case float accelerated when (accelerated > 110):
+                case float accelerated when accelerated > 110:
                 {
-                    Projectile.velocity = (player.Center - Projectile.Center).SafeNormalize(Vector2.Zero) * Leecharang.ShootSpeed * 1.25f;
+                    Projectile.velocity = (player.Center - Projectile.Center).SafeNormalize(Vector2.Zero) *
+                                          Leecharang.ShootSpeed * 1.25f;
                     break;
                 }
             }
-            if (Projectile.Hitbox.Intersects(player.Hitbox))
-            {
-                Projectile.Kill();
-            }
+
+            if (Projectile.Hitbox.Intersects(player.Hitbox)) Projectile.Kill();
         }
     }
 
@@ -83,35 +86,25 @@ public class LeecharangProjectile : ModdedFriendlyProjectile
     {
         target.AddBuff(ModContent.BuffType<LeecharangBleed>(), Leecharang.DebuffDuration);
         if (target.GetGlobalNPC<ParadoxNPC>().LeecharangBleedStacks < Leecharang.DebuffMaxStacks)
-        {
             target.GetGlobalNPC<ParadoxNPC>().LeecharangBleedStacks++;
-        }
-        if (!Returning)
-        {
-            Return(true);
-        }
+        if (!Returning) Return(true);
     }
 
     public override bool OnTileCollide(Vector2 oldVelocity)
     {
-        if (!Returning)
-        {
-            Return(true);
-        }
+        if (!Returning) Return(true);
         return false;
     }
+
     /// <summary>
-    /// Puts the Boomerang into a returning state.
+    ///     Puts the Boomerang into a returning state.
     /// </summary>
     /// <param name="HitCollision">
-    /// Whether Return was called by hitting a npc or tile. Set to true to make it return at full speed instantly.
+    ///     Whether Return was called by hitting a npc or tile. Set to true to make it return at full speed instantly.
     /// </param>
     private void Return(bool HitCollision)
     {
-        if (HitCollision)
-        {
-            AITimer1 = 111;
-        }
+        if (HitCollision) AITimer1 = 111;
         Projectile.tileCollide = false;
         Projectile.penetrate = -1;
         Returning = true;
